Added reclassifying To Dos and Next Actions

+++ GettingThingsDone/NextAction+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  NextAction+CoreDataProperties.swift
 //  GettingThingsDone
 //
-//  Created by Dylan Bruschi on 3/14/17.
+//  Created by Dylan Bruschi on 3/15/17.
 //  Copyright © 2017 Dylan Bruschi. All rights reserved.
 //
 
@@ -20,8 +20,8 @@ extension NextAction {
     @NSManaged public var details: String?
     @NSManaged public var duedate: NSDate?
     @NSManaged public var name: String?
-    @NSManaged public var priority: Int16
-    @NSManaged public var processingtime: Int32
+    @NSManaged public var priority: Float
+    @NSManaged public var processingtime: Float
     @NSManaged public var id: String?
     @NSManaged public var project: Project?
     @NSManaged public var topic: Topic?
